requirements.txt
airtable-python-wrapper==0.8.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0
idna==2.6
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wincertstore==0.2
